requirements.txt
instaloader==4.13.1
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.23.0
TikTokApi==6.5.2
pyTelegramBotAPI
instaloader
TikTokApi
